gstlal_demodulate:  Remove dependence of output on start time of pipeline. gstlal_compute_strain:  Added DARM line coherence gating to kappa_tst calculation.

+++ gstlal-calibration/gst/lal/gstlal_demodulate.h
@@ -73,7 +73,7 @@ struct _GSTLALDemodulate {
         gboolean need_discont;
 
 	/* properties */
-	double line_frequency;
+	int line_frequency; /* in centihertz */
 };
